add browser logging
firefox logging may not actually work due to
https://github.com/SeleniumHQ/selenium/issues/2698, however upgrading
selenium to get that fix (selenium >= 3.0.1) is a prohibitively big
change due to differences in how firefox is handled.

we should probably upgrade to 3.0.1 anyway because upgrading is good,
but the changes to how firefox runs (marionette instead of natively)
means that's not a straightforward change if we stick with firefox -
a bunch of our tests break due to changes in how page changes interact
with element finding etc. or we switch to chrome and stuff *might* change