Allow Selenium to check new radios/checkboxes
The new checkboxes and radio buttons hide the native elements by using
`opacity: 0`. Selenium considers elements that are completely
transparent to not be on the page. It will refuse to click elements that
it doesn’t consider to be on the page.

The workaround is to directly set the `checked` attribute of the
`input`s, rather than do so (as a user would) by clicking.

I think this is fine because we’re not testing the UI, we’re testing the
implication of (for example) choosing a test API key.