Disable test_all_the_things
This code wasn’t being run as part of our normal deployment pipeline. It
was only being run when updating the functional tests. But since it’s
out of date (not consistent with the app any more) they are blocking
updating the functional tests when the preview tests stop working.

This is a quick and dirty way of unblocking things.

Should review what of this code is still needed at a later date.