Add flake8 linting to project The GDS Way™[1] recommends using Flake8 to lint Python projects.
This commit takes the Flake8 config from Digital Marketplace API[2] and
removes the bits we don’t need.

It changes the `max_complexity` setting to 14, which is the most complex
code we have in this repo currently (we shouldn’t be writing code _more_
complex than what we already have).

This commit also fixes the errors found by Flake8:
```
./tests/test_utils.py:14:1: F401 'selenium.webdriver.support.ui.WebDriverWait' imported but unused
./tests/pages/__init__.py:1:1: F403 'from .pages import *' used; unable to detect undefined names
./tests/pages/__init__.py:1:1: F401 '.pages.*' imported but unused
./tests/pages/pages.py:13:1: F401 'tests.pages.element.ServiceOrgTypeElement' imported but unused
```

1. https://gds-way.cloudapps.digital/manuals/programming-languages/python/linting.html#how-to-use-flake8
2. https://github.com/alphagov/digitalmarketplace-api/blob/d5ab8afef4a0472f9d266d94ab4ffe1333a9aaad/.flake8