functional tests now have two preset status sets to check against
You're encouraged to pass in `NotificationStatuses.SENT` or
`NotificationStatuses.DELIVERED`.

We've seen problems with our provider tests (which go to a twilio number) failing because of permanent failures.

If we get a delivery receipt, we know we got through to the providers - so lets just check for that rather than specifically expecting success