automatically read the right config based on the tests being run
previously we were relying on the ENVIRONMENT env var to choose what
vars to source - but document download has env vars separate.

This solution means that if you are running any tests from the
preview_and_dev folder, it'll load the preview config. If you are
running any from the document_download folder (whether in the same
test run or not) it'll load the doc download config too - but if not,
it'll not try and access them.