Add dev/preview inbound sms api test
This test mimics the MMG inbound sms api call. When one of our providers
receives a text to one of our inbound inboxes, they send us a POST
containing the user's number, the content, and a couple of other bits
and bobs. This test mocks out the provider and hand-crafts a payload to
send directly to the endpoint.

Then, we do an API call to ensure that we can see the content in the
returned list.