remove setAttribute hack and add click to stale methods
setAttribute was added in 125fe3749 to allow us to click radio inputs
that aren't visible. Due to either changes in selenium or govuk frontend
toolkit, this isn't necessary now.

The AntiStale element overrides getattr, so that calls like `el.text`,
which make a live call to the DOM, don't fail. However, `el.click` just
returns the function pointer - it's the call itself that makes the DOM
check. So override `click` too. Also refactor to reduce code usage
through lots and lots of lambdas