update doc download test to send a "real" pdf
previously it was sending text - but the latest version of doc dl on
preview only accepts PDFs. it does this by checking the file we send
for the magic pdf bytes, so make sure they're present in the data.

Also, change the assert from a raise_for_status to a check of the
response json so that we can see errors