fix bug where template type radio doesn't select
for some reason, selenium doesn't think the radio button is visible,
even when the page has finished loading and nothing's changing on the
page. if we check for only presence and not visibility, then it tries
to click the radio before it's selectable. so to counter this, we wait
for the continue button to appear (which selenium always recognises as
visible), and then click the radio button.