don't check notification status when retrieving 2fa codes
we were waiting for the notification to enter "sending" status before
trying to log in with it - that's unnecessary and slows down tests and
makes them less reliable. status is now optional.

also clean up the retry logic: don't retry the get_notification call,
because we'll then retry that as well as the outer call to login which
also has retry logic. Rely on that - replace the pytest.fail with a
retry so that we still retry somewhere.