Make user pages work with new permissions UI
Makes the locators more specific and checks for
the new UI before changing permissions in case the
checkboxes for them need to be revealed.

Also removes `show_folder_permission_checkboxes`
method from the `InviteUserPage` class. Having it
meant the tests wouldn't work with the present
state of the codebase.

This also includes a change to the previous commit
where a locator for the 'Choose folders' button
references the aria-controls attribute instead of
the '.button-secondary' class. The intention is to
prevent any future issues with using a class that
could be applied to other elements in the page.