Simplify the update email reply-to address test
Also increase wait time for delivery receipt.

We simplified the update email reply-to address test so that
it doesn't take too long. The lines we deleted are already covered
either by other functional tests or by unit tests, or both.

We increased maximum wait time to 120 seconds, because
the longest wait we saw in the notification history was
between 70 and 80 seconds, and we wanted to have a breathing space
ensuring that tests do not fail because of a slow queue.