Make template link XPath more robust
`text()` returns a nodelist. When the child is just text, this is a list of 1 element. When the child is interrupted, for example if it contains some HTML elements (or even a comment tag) it will return multiple text nodes and `contains()` only looks at the first. Using `.` instead of `text()` fixes this problem. Using `normalize-space` truncates runs of whitespace, which also makes things more robust. 

Better explanation here: https://sqa.stackexchange.com/a/36921