Change Dockerfile to use Python 3.5 and remove phantomjs
Removed phantomjs since it is deprecated and we weren't using it
(along with node and npm which we needed to make it work).

Trying to install Chrome in the Dockerfile was giving errors, so
switched to only install Firefox and to use it as the default. It is
still possible to use Chrome when running the tests locally.