make sure we wait for email validation in reply to flow
click_save checks for a button using `(By.CLASS_NAME, 'button')`. My
suspicion is that the email validation holding page hasn't loaded, and
selenium grabs the previous "add" button on the actual input form page.
It then clicks that button a second time (thinking it's already sat
through validation and then hit the continue button), and then moves
straight on to checking what the page is.

By moving from the generic click_save to an xpath that specifies the
button text to expect, we should solve this by not double-clicking the
add button.