Set chrome driver window size to allow tests to pass on both CIs
We set the window size to be large because without doing so when
using headless the window is too small for some of our tests and
some content we assert on will not appear, making our tests fail.

In particular, the test that caused the failure was
`test_registration_and_invite_flow`, where the string `(default)` would
be cut off for smaller window sizes.

All the tests now pass on both Concourse and Jenkins using Chrome
headless.