Update Python to 3.7 when running in Docker
We can’t upgrade to the newest version of Pytest while we’re running on
Python 3.4. We’ll be running on Python 3.7 once we move to Concourse,
which uses the other Dockerfile.

We've slimmed down our dockerfile to match the concourse one just to get
the tests working as quickly as possible since the move to a newer linux
base image. We've removed support for things like phantomjs and firefox
in doing so.