make virus scan checks actually check that virus scan has run
do this by checking that the status of the notification goes to
received/virus_scan_failed respectively, rather than just
pending_virus_check.

also, fix a reference to refer to virus_scan rather than delivered for a
bad letter to help us debug, and also mark all the tests as antivirus so
that we can refer to them nicely from concourse.

Note that the antivirus mark is not an exhaustive list of every test
that goes through the antivirus flow, but it at least covers a few
cases and that felt like enough.