Update locators for the new button style
We are now switching notifications-admin to use the govuk-frontend style
of buttons, which means that most of the button locators need to be
updated. We can't find the buttons using the new class name of
`govuk-button` because the cookie banner buttons also have this class
and they are present in the HTML even when they aren't visible on the
page. Finding the button using the CSS selector of `main
button.govuk-button` works in most cases since the cookie banner buttons
are not in the `main` element.