Reorder code in _select_template_type
Before, we were getting both the continue button and radio element
from the create a template page, then clicking both in turn. This
changes the order to get each element just before we click it, to
decrease the risk that the element is no longer visible on the page when
we need it.