Add smoke test for pre compiled letter upload via the API
This will help us catch if we break something about in production.
Note, it has been done using a test API key as we don't want to send a
real letter ever as this will cost us money.

This is pretty much a copy and paste of the code for the preview and dev
version of this test.