Stop using research mode service
We want to be able to test that templated letters get created, which is
not possible in research mode, so we will turn off research mode for the
functional test service.

This updates the sql data dump and an assertion about a letter state to account
for the service now not being in research mode.