Add test to see if templated letter is created
To see if a templated letter is created we check if a PDF file for it
has been created in S3. It's not possible to do this by checking
billable_units for the letter or something similar, because that does
not get returned from notifications-api and we don't have access to the
database to check.

To check if the letter is in S3 we use the client method of `get_pdf_for_letter`,
which raises an exception if the letter is not found.