Change how message item status element is found
If virus scanning doesn't happen, the 'updated_at'
field isn't shown. This means locally the tests
can't find the status element because they look
for the 6th field and there are only 5.

Whatever happens to the message, the test is
asserting against the status, not its presence.

This changes the CSS selector to find the last
field in the message details, making it work
however many fields are present.