Open 2nd message log to check view letter link
This precompiled letter test contains a final
assertion that the API page message log for it
will not contain a 'View letter' link because it
failed the virus scan:

`tests/functional/preview_and_dev/test_seeded_user.py::test_view_precompiled_letter_message_log_virus_scan_failed`

This assertion tests this by:
1. looking for the first 'View letter' link in the
   page
2. asserting it isn't for the message log related
   to the letter that failed by checking its
   `href`

This is also explained in
https://github.com/alphagov/notifications-functional-tests/commit/a38d182b38edd8d558dc199b0e8bdb2c57cb32a5.

This works on preview because it can find all the
'View letter' links in the page, despite them
being in closed `<details>` tags. This isn't the
case when run locally.

This opens the second `<details>` tag so the 'View
letter' link for that message log is visible. This
lets it be found locally and doesn't effect the
test when run on other environments.

This came out of work done on previous problems
with tests that assert things in the API page:

https://github.com/alphagov/notifications-functional-tests/pull/339