Open all API log entries, not just the first two
Only opening the first two entries assumes a lot
about what will be in the page. This opens all of
them to make all of their contents visible,
similar to what the Firefox/Webdriver combo on
preview sees.

Adds a `__len__` magic method to the
AntiStaleElementList to allow the number of
elements returned to be queried.