Add mixin for pages with sticky navs
For pages with a sticky nav, to let tests scroll
the page so that the element sent is not obscured
by the sticky nav.

This is driven by the assumption that webdriver's
built-in methods for determining an element can be
interacted with cannot know when scrolling stops
an element from being obscured by the sticky nav.

For example, the webdriver spec says that a click
action is preceeded by scrolling the targeted
element into view:

https://www.w3.org/TR/webdriver/#element-click

The problem here is that an element's position can
be within the bounds of the window frame but still
obscured by the sticky nav.

This method should let tests be sure an element in
this kind of page is not obscured by a sticky nav
before attempting to click it.