Make assertion of document download service name generic
We wish to run this test in production. It already runs in preview and
staging. However, it would currently fail in production because the name
of the service in production is "Functional Tests - CSV & Notify API"
whereas in preview and staging it's just called "Functional Tests". This
commit will make it pass in both.

I could make our functional tests smarter by taking it in as a variable
for the service name but I don't think it's worth this effort when I can
fix it quickly like this. Unit tests will already cover the service name
being shown and the assertion used is still pretty good here.