Fix flakey test due toggling of log entry
Previously the "retry_call" function would lead to the list of log
entries being in an undefined state: each invocation of the method
to "get_status_from_message" would alternately expand and collapse
the first log entry. Consequently, attempting to get the first link
in "get_view_letter_link" could exhibit strange behaviour, such as
returning the link for the second log entry by mistake.

This modifies all the methods that interact with the log entries to
make sure the entry is in a collapsed state when they return. Since
the ability to click on the log entries proves all the elements are
on the page, the calls to "wait_for_element" were redundant, and we
can use normal "get_element(s)" calls instead.