Add missing info about provider delivery tests
This modifies the introduction to clearly distinguish between the
types of test in this repo, as opposed to the technologies used,
which are more incidental. As before, there's a discrepancy between
the scripts and the way the tests are organised, so this reverts to
manual "pytest" commands for running them.