Move remaining test script into Makefile
This follows the same approach we have in other repos. Note that
the test script is not used by CI, which has its own macro.

I've restricted the new make task to run tests against the local
environment (or potentially Preview). We don't normally need to
run tests against Staging and Prod from a local machine, so it
makes sense to just document this in the README if it's needed.

I've kept the "-v" flag for pytest, since it's useful to know which
test is being executed, as each of them are slow to run. I've not
kept the status check on the Admin app, as it's unclear why it's
necessary - it's just one part of the required environment.