Build Docker image using make task
This mirrors the approach we're taking in other repos [1][2], which
ensures the setup of the image is consistent with local setup.

I've added a .dockerignore file to compensate for using "COPY . ."
to make all files available in the image. While this image is only
used on CI and a local developer is unlikely to push it anywhere, we
should avoid baking files with secrets into any images.

[1]: https://github.com/alphagov/notifications-ruby-client/pull/122/commits/94858af5bbbe8ddf60ac9c544fcd96cb7f058cc3
[2]: https://github.com/alphagov/notifications-template-preview/pull/556