WIP test that alert gets published on govuk-alerts page for preview when we send the alert. It does not work at present
I think this is not happening because our preview service is trial
mode, and trial mode services broadcasts do not land on govuk-alerts.

If it was a live service, it would need the approver to
use a yubikey, and we have not figured out signing in with yubikey
for functional tests.

Does this mean that this piece of work is blocked or currently
untenable?