Fix not being able to run broadcast tests locally
running the functional tests locally is a useful way to check my
local environment is working properly. these tests are also in the
"preview_and_dev" directory, so they should work locally.

this fixes the urls for local testing and also an inconsistency in
the service settings compared to preview: the tests check an alert
shows up on /alerts, but this won't happen for test alerts.