Update link text for downloading a doc download file
I went for a partial text match rather than updating the link
to include (0.1KB) as that feels like detail that the functional
tests don't need to care about, plus I know it will shortly change
to also add the file type as well. However I don't feel strongly
about this choice and can instead change it to be more precise if
needed.