add new test for email confirmation flow
in the un-auth'd flow we use requests to just download the file and
check the contents. If you use selenium to navigate chrome to that page,
it'll just load an iframe with the pdf renderer inside so you cant
assert conents.

however, with the auth'd flow it relies on the cookie being set to
access the file. requests doesn't have the cookie (only the chrome
instances does) and i can't figure out how to get that cookie out, since
driver.get_cookies() doesn't return it (it does return the other cookies
on path '/').

so, to solve this, instead of getting a pdf, get a csv instead - when
chrome renders that, it just shows the text with a minimal html
wrapping, which is easy to assert on.