test document download via notify api
we don't really have current plans for the document download api to be
accessed directly and publicly, so it makes sense to test the entire end
to end flow. this can ensure we're using the flags passed in to
prepare_upload correctly.

this also means we can get rid of some additional configuration from the
settings as we had to have the internal doc dl api key stored in the
config.