check both chrome window and download folder for downloaded doc
for me, locally, when i open a text file from document download, it
opens within the chrome browser. However, when i run the same tests
against preview, the file is downloaded. I'm unsure as to what the
differences between localhost and notify.works are but rather than dig
in to that, it's easier to make the functional tests a bit more
resilient and cater for both situations. If, after clicking the link,
the current_url is that of the link, then it's rendered in browser, and
we should check as before. If the URL hasn't changed, then we've
downloaded the file. Look for it in the downloads folder and wait up to
10 seconds for the download to finish.

Use a temporary folder (tmp_path_factory is a pytest provided fixture
that creates a subfolder for each xdist under an arbitrary tmp folder
and cleans it up afterwards) so we don't pollute any local folders with
test files. With thanks to https://stackoverflow.com/a/40656336 for the
chrome settings required.