create separate preview and staging doc dl functional tests
we run smoke tests every ten minutes on prod - these can't use the
seeded config as that is full of environment variables that are only
used on preview on the full suite of tests.

So, follow the lead of hwo we've strucutred notify admin tests, and
split into two folders - preview_and_dev and staging_and_live.

The preview_and_dev folder is just the old files unchanged.

The staging_and_live folder only has a single test to check a single
document being sent. It uses the email confirmation flow just so we
don't have to rewrite it later. (It's just the same test copy pasted).

we also remove the `@pytest.mark.antivirus` because that is a trigger to
run the tests when antivirus merges, and should only apply to
preview/staging methods.