Add required DB fixtures for functional tests for staging
The fixtures only get the functional tests passing, not
the document download functional tests, nor the functional
provider tests. These will need to be added later (whether they
go in this same file or in a new file hasn't been decided yet).

We decided that although the secrets and API keys are already
hashed, we might as well set a good precedent and not store them
in the public repo, but instead pull them in from the creds repo.

This gives us the possibility of having similar scripts for preview
and production. This would be useful if we ever accidentally
lost our fixtures from our preview or production databases, but
is probably not urgent as the chance of this is low.

The reason we've bothered doing this for staging, as the chance of
losing our fixtures that already exist is low, is that we are hoping
to replace our staging database with a new one that is a redacted
version of production. That redacted version of production won't
have these exact fixtures in, nor would they work anyway because
we redact the things like passwords and API keys too.