move test_change_service_name to non-seeded tests
if this test fails for whatever reason, it might leave the service in an
unusual state (with the wrong service name). lots of tests check the
service name to assert they're logged in to the correct account, and
thus, will fail unless we manually rename the service back.

moving this test from the seeded service to the test registration
service means we can not worry as much if the test fails, as no future
test runs will look at this service.