Remove references to 'jenkins' from all variables
We no longer use jenkins so we can remove all references to it.

Rather than replacing 'jenkins' with 'ci' or 'concourse', lets
just remove it because it isn't important (I'd like us to later
explore if we even need to include a build ID anywhere as I don't
think we are getting any value from it).

I've also changed all the environment variable names that reference
jenkins. The corresponding credentials repo change will need to be
merged before this will.