Fix environment variable name
In the creds repo, these are named 'TEST' not 'TESTS'.

It looks like there is some inconsistency generally on whether
to include the S which I'm not going to try and solve as part of
this PR because it has broken our deployment pipeline so am
going for speed. We can look at fixing the inconsistency later.