Remove passing of `BUILD_ID` by environment variable
My rough recollection/assumption is that we used to pass a jenkins
build ID into the functional tests whenever we did a deployment. We
stopped doing that when we moved to concourse so we no longer
appear to be passing `BUILD_ID` as an environment variable anymore
(I've searched the creds repo and also the concourse pipelines).

I looked into whether we can just remove the `build_id` completely,
ie from the notify templates themselves if we are just passing
'No build id' all the time. However it appears that we do indeed
use that placeholder, for example we use it to pass in a link to
document download files in:
https://github.com/alphagov/notifications-functional-tests/blob/main/tests/document_download/preview_and_dev/test_document_download.py#L24

Therefore we can't remove the build_id completely, but I think it
is still nice to remove a couple of lines of code, in particular
for the complexity when you have to wonder if the build_id is
being provided by the environment ever or not.