Remove HTTP proxy support
We aren't using this anymore. It was likely needed when we used
jenkins and used an egress proxy but we no longer do. We can
probably remove just the couple of lines of code. If we find out we
need it again, we can add it back in.

Note, given how small an amount of code this is, I'm very open to
keeping it in, as it isn't causing a particular increase in
complexity to our code.