Add Pytest marker for the broadcast tests
We have Pytest markers so that we can run the antivirus and
template-preview tests. This adds a new marker, broadcast, which we will
use to _avoid_ running the broadcast tests for the ECS apps using
`pytest -m "not broadcast"`. We are not creating the broadcast worker in ECS,
so the broadcast tests will always fail and add extra running time.