Run seeded user tests with unique accounts
The slowest group of tests in our functional test suite at the moment
is the xdist group 'seeded-user'. These tests all run using a
pre-defined user account in a specific organisation and service, with
specific permissions. The tests don't necessarily step on each other's
toes that much, but our app enforces only one user session can be active
at a time. If two of the seeded-user tests run at the same time, one of
the sessions will be killed by the other.

To deal with this, let's instead use a new private API endpoint to
create a 'seeded' user for each test that needs one. This will let all
of those tests run in parallel and give us a much faster runtime for the
full suite of tests.

At startup, we analyse all of the tests that have the
'login_seeded_user' fixture, and post a request to the API endpoint to
make sure that users exist for all of those tests. We then rewrite the
`get_email_and_password` helper to return the unique per-test email
address. We only execute this flow on one of the workers using a
temporary file to guarantee only a single execution.

There are a few tests that read/write templates and other central
objects at the same time. Where tests do that, we make sure to either
create each template with a unique name, or use an arbitrary locking
method to ensure only one test can be inside a given block of code at a
time. The latter is most relevant for letter template tests, as letter
templates are all created with the name 'Untitled letter template' and
need to be made unique in multiple steps.

For a locking mechanism we can use a simple file-based lock, as the test
run is not executed across multiple machines.