sign out if we're already signed in when trying to sign in
many tests sign in at the beginning. if they then fail, they'll be
retried on concourse but the user will stil be signed in. the first
thing the second run of the test does is try and sign in again, which
will throw an exception when it tries to go to the sign_in page cos
it'll redirect to /accounts-and-services which the functional tests
don't expect.

so instead, catch this error, and try to sign out before continuing