install chromedriver that matches the version of chrome
google-chrome-stable currently returns 120.0.6099.199. however, for
reasons i can't figure out, the chromedriver website[^1] only tops out
at version 115, otherwise it points you to the chromedriver for testing
page. (Note that 115 is six months old. 120 is only a week or so old).
You need to have a matching version of google-chrome and chromedriver,
so we need to either test against an old version of chrome or make sure
we get a newer version of chromedriver. Pinning to an old version seems
likely to fall behind pretty quick, so instead get the latest version of
chromedriver as referred to in last-known-good-versions.json on the
chrome github[^2].

The eagle eyed may notice that there's still a difference between the
stable as reported by googlechromelabs.github.io and the stable as
returned in the debian package (120.0.6099.199 vs .109). The correct
answer is probably to also download google chrome from source from the
github link (so we know we have a totally matching version). However,
for now while these two work I'm okay to leave it as is to avoid us
making this dockerfile even more handwritten.

Also note that the unzip had to change slightly as the zip file layout
changed from the old chromedriver website to the new github repo link

[^1]: https://chromedriver.chromium.org/downloads
[^2]: https://github.com/GoogleChromeLabs/chrome-for-testing#json-api-endpoints