Force all requests to PaaS by default (not ECS)
This mimics the equivalent for ECS
(https://github.com/alphagov/notifications-functional-tests/pull/459/files)
apart from we make it the default so that we don't need to go
and update all our PaaS functional test pipelines to include
a new environment variable.

Note, on ECS once a request enters ECS, all inter app requests also
stay in ECS. This is NOT true for the PaaS though. For example
if we send a request to the admin app in PaaS and the admin app
needs to query the API to get some data, then the request to the API
goes back out through cloudfront. The consequence of this is that
this change doesn't enforce all requests will stay in the PaaS
because inter app requests may go back out to cloudfront and be
at the mercy of cloudfront percentage based traffic routing.

We have a few options:

1. Tolerate the above behaviour, knowing that at least the initial
  request will go to the PaaS. Also not all journeys will be
  affected because they don't make interapp requests, for example
  sending an email via our API doesn't make an interapp requests.
2. Change the behaviour in the PaaS so that all inter app requests
  are forced to go to the PaaS. We can do this by changing
  their URLs to be the PaaS provided ones (like
  app-name-production.cloudapps.digital).
3. Write some python code in our apps so that inter app requests
   will inspect the HTTP headers on the request and then propogate
   that header onwards for any inter app requests it makes.