Add test for inbound SMS callbacks
This test relies on a third party request bin service[^1]. We configure
the Functional Tests service on Notify to send inbound SMS callbacks to
a source configured on Pipedream. We can then use the Pipedream API to
verify that the correct callback was triggered when an inbound SMS was
received.

n.b. the bearer token added in local.sql is "0000000000". It's signed
before it's inserted into the db for historical reasons, but there's no
need to keep this value secret and so it's fine for it to live in this
repo.

[^1]: https://pipedream.com