Use Debian-packaged chromium & chromium-driver
We currently install the latest version of Chrome from Google’s apt
repo, and the latest version of Chromedriver from chrome-for-testing.
This is brittle, because it’s not guaranteed that we’ll end up with
matching versions of Chrome and Chromedriver.

We should use Debian’s own chromium and chromium-driver packages
instead.

Selenium seems to be able to find Chromium automatically (even though it
lives in a different path than Chrome), so we don't need to make any
changes to our test config.

We will continue pinning Chromium v126 for now, because v127 and later
are incompatible with our functional tests.