Treat absence of banner as ‘0’ unsubscribe requests
The banner only shows if there are some requests in the last 7 days.

At the moment the tests treat absence of the banner as a failure.

This will be a problem if the tests haven’t run in a while and all the
previous unsubscribe requests have been cleaned up (they automatically
get deleted after 7 days).

This commit changes the tests to treat absence of the banner as
equivalent to ‘0’ requests (which is true). This means they can continue
to the next step which is creating the first request.