set reruns to zero by default
remove reruns - we're worried that this will lead to inconsistent
state, particularly in tests that start with signing up a new user. If
that fails, then we don't see anything, and can end up with lots of
questions such as:

* What error did we get?
* Why did it fail
* What page is it on now, and does the unit test assume that it's on a
  particular page at the start of the function?
* which artifacts of the functional tests (eg user accounts, services,
  templates) have been created?
* If we retry will it try and recreate those when re-running, will
  Notify handle that or throw an error?

I've decided that I'd rather understand our failures rather than try and
just smooth over them

If it means we need to bump up some individual function retries on
things that sometimes time out (eg letter creation), that is much more
preferable even if it might mean a little more churn in editing this
repo

note that we're just setting the reruns to zero rather than removing the
plugin so that we can safely remove the `--reruns 0` flags we have in a
few places already without any downtime. then we can remove the plugin
entirely once nowhere references reruns