log all selenium activity and print out last N actions on error
Selenium has an event listener base class[^1] - create a subclass of
this that logs every time we perform basically any selenium event.

Then, if there's an error, we print out the last 20 events the selenium
web driver performed

This will help make debugging easier - We already have some idea of
"What page was the test viewing when it failed" and "what was it trying
to do at time of failure", but if we realise it's on the wrong page,
it's quite hard for us to tell how it got into that state - this commit
should help with debugging that.

[^1]: https://selenium-python.readthedocs.io/api.html#module-selenium.webdriver.support.abstract_event_listener