make 2fa sms code logic more robust
i have a theory that it was finding a race condition where the login
would fail, but the bit of code where it tries to find the error-message
would fail to find the element _before_ the page had reloaded and error
message could be rendered.

my gut says that anywhere we say "driver.find_element" without using a
`WebDriverWait` could be suceptible to this too, unless we're confident
the page has already finished loading.

change the login so that if it cant find the error message, it checks
to make sure we're on a different page (which would indicate that we
logged in succesfully)