check for /verify and /two-factor-sms on verify page
verify for new user registration
two-factor-sms for returning user sign-in

prev we were just checking for "does url not equal /verify" so got false
positives on the user sign-in flow.

also parse url and check path for equality

rather than checking path in a whole URL, split the path.
this means we can be robust (and not accidentally match on URLs that
look like `/verify-email` for example), while not getting false
negatives if the URL includes, eg, query strings