double timeout for csv uploads in smoke tests
sometimes, the smoke tests happen at the same time as a large upload of
jobs, and the notifications are stuck in a queue behind that. if we wait
twice as long for notifications to move to sending, that gives us an
extra buffer for these cases.

when users upload jobs, our assumption is that it's not actually as
important if that notification sends within 10 seconds (when compared to
one-offs or api messages which may be part of synchronous user
interactions like two factor codes). so it's okay if it takes a bit
longer