Add random phone numbers
In the config for functional test there is only one number that all tests cases use when registering the users. There are multiple tests run in parallel and causing the issue. Ofcom provides test numbers from 07700 900000 to 900999 and these numbers are not allocated to anyone (https://www.ofcom.org.uk/phones-and-broadband/phone-numbers/numbers-for-drama/). Added logic to generate the random number from this series.