test_can_add_and_update_reply_to: use proper skip mechanism, make skipping overridable
we need to move away from making random decisions based on sniffing
the env name.

specifically in this situation we're very soon going to run
test_can_add_and_update_reply_to against staging instead of preview,
so make the decision of whether to run this or not explicitly
settable, otherwise falling back to a default based on the env name
in order to retain existing behaviour.

also use the proper pytest skip mechanism instead of just "passing"
the test.